src/requirements.txt
django
bilibili-api-python
